Create BuildBasicStructure.yml
When it is before workflow, the prerequisites to the joining constructs, are better to have ready and aware for planning forward on CI(continuous improvement) model. So, remember, be ready with what is all  involved in the construct of the build. Keeping Workflow in the work path of you the project is important to have, also in the AddPath of it.